Refactor calendar utility class into calendar service class
